Add test with links pointing nowhere
