+++ ldes/shared/src/test/scala/run/cosy/ld/JenaWebTest.scala
@@ -7,6 +7,6 @@ class JenaFoafWebTest extends FoafWebTest[R]()
 
 class JenaLdesSimpleWebTest extends LdesSimpleWebTest[R]()
 
-class JenaLdesCircularWebTest extends LdesCircularWebTest[R]()
+class JenaLdesBrokenWebTest extends LdesBrokenWebTest[R]()
